+++ app/assets/javascripts/admin/application.js
@@ -13,4 +13,4 @@
 //= require jquery
 //= require jquery_ujs
 //= require turbolinks
-//= require_tree .
+//= require ./locations-directory.js
