+++ app/assets/javascripts/admin/locations-directory.js
