+++ app/assets/stylesheets/admin/components/_location-details-table.scss
