+++ app/assets/stylesheets/admin/components/_pagination-atoz.scss
