+++ app/assets/stylesheets/admin/layout/_location-status.scss
