--- app/assets/stylesheets/application.scss
@@ -1,10 +1,0 @@
-@import "govuk_admin_template";
-
-@import "lib/variables";
-@import "helpers/contain-floats";
-
-@import "components/location-details-table";
-@import "components/pagination-atoz";
-
-@import "layout/filter-form";
-@import "layout/location-status";
