+++ app/controllers/admin/base_controller.rb
@@ -1,0 +1,10 @@
+module Admin
+  class BaseController < ApplicationController
+    layout 'admin/application'
+
+    include Pundit
+    include GDS::SSO::ControllerMethods
+
+    before_action :require_signin_permission!
+  end
+end
