+++ app/views/admin/locations/_directory_filter.html.erb
@@ -1,4 +1,4 @@
-<%= form_tag locations_path, method: 'get', class: 'l-filter-form' do %>
+<%= form_tag admin_locations_path, method: 'get', class: 'l-filter-form' do %>
   <ul class="pagination-atoz">
     <% (?A..?Z).each do |letter| %>
       <li class="pagination-atoz__letter t-pagination__letter">
