+++ app/views/admin/locations/index.html.erb
