+++ app/views/layouts/admin/application.html.erb
@@ -1,5 +1,5 @@
 <% content_for :head do %>
-  <%= stylesheet_link_tag 'application', media: 'all' %>
+  <%= stylesheet_link_tag 'admin/application', media: 'all' %>
 <% end %>
 
 <% content_for :navbar_right do %>
@@ -28,7 +28,7 @@
 <% end %>
 
 <% content_for :body_end do %>
-  <%= javascript_include_tag 'application' %>
+  <%= javascript_include_tag 'admin/application' %>
 <% end %>
 
 <%= render template: 'layouts/govuk_admin_template' %>
