+++ config/routes.rb
@@ -1,7 +1,11 @@
 Rails.application.routes.draw do
-  resources :locations, only: [:index, :update]
 
-  root 'locations#index'
+  namespace :admin do
+    resources :locations, only: [:index, :update]
+    root 'locations#index'
+  end
+
+  root 'admin/locations#index'
 
   mount GovukAdminTemplate::Engine, at: '/style-guide'
 end
